Remove old functionality from profile controller.
In the past you could visit a URL 'site.com/profile/-/function'
that would sign you on and take you to that function, but this
has not been used for a long time. Administrative emails directing
you to do something with your profile insert your profile URL name
into the path, not a dash.