Sign on menu button returns you to any page you clicked it from.
Clicking sign on would take you to whatever page was specified in
the config as the default page once signed on. This was lousy
behaviour, as often you sign on from a specific page to get see
additional functionality. Now sign on will return you to wherever
you were. Only if no page is specified will you go to the default
signed on page.