Fixes signon URLs across the CMS
I'm not sure how this issue went unnoticed for so long. Perhaps CI
interprets uri_string() differently on different versions of PHP,
because that's the only thing that's changed to bring this issue
into the spotlight. Either way, I was sloppy with generating the
signon URL, with potential to duplicate 'signon'. This fixes that.